Improve data type coverage for Image.
Added type coverage for get_access() for cl_int4, cl_uint4, cl_half4
Added type coverage for Image<N, AllocatorT> for AllocatorT - not
default
Added type coverage for constructors for:
 image_channel_order = (image_channel_order::a, r, rx, rg, rgx, ra, rgb, rgbx, argb, intensity, luminance, abgr)
 image_channel_type = (image_channel_type::snorm_int8, snorm_int16, unorm_short_565, unorm_short_555, unorm_int_101010)

Remaining gap (test coverage): other accessMode for get_access(),
                               constructor image(cl_mem , const context &, event )
